ai_tools/1/requirements.txt
pillow
torch
torchvision
scikit-image
pycryptodome